renamed keyvalue to kv and moved to project top. Using better CmpFold function for search
