+++ kv/storage_test.go
@@ -1,4 +1,4 @@
-package keyvalue
+package kv
 
 import (
 	"github.com/stretchr/testify/require"
